Modified how slice order is handled in PAR files to fix problems with files in which slices and volumes are intermixed
dicm_hdr.m: creates a slice order vector across all volumes instead of assuming that the same order is used for all slices and that volumes (dynamics) are in ascending order. For function scans, the dynamic number and image type (magnitude or phase) is used
 to reorder. For diffusion images, the b value and diffusion orientation number is used. Note that the SliceOrder field is now created for all PAR/dicom header conversion, even if all slices and volumes are in ascending order (in which case the slice order does not change)
dicm2nii.m: how the slice order vector is applied depends on whether is it si as long as the total number of slices (PAR) or the number of slices in a one volume.