Change default bids gui width to 800
The default gui width when converting to bids and setting skips and comment is often too narrow to see all the columns. This isn't so bad, however one of our studies requires a lot of manual intervention and it can be quite tedious to always extend the window manually. If there is a reason why the default width was so small, then feel free to reject this pull. Otherwise, what do you think of upping it to 800?